+++ packages/nuxt-base-layer/src/components/Chip/AutoDelete.vue
@@ -1,15 +1,22 @@
 <template>
   <chip-date
-    v-if="props.value"
-    :value="props.value"
+    :value="(props.value === 'never' || !apiSettings.isProfessionalLicense) ? null : props.value"
     icon="mdi-delete-clock"
+    label="delete"
     tooltip-prefix-text="Automatic deletion: "
     relative="future"
+    :color="isSoon ? 'error' : undefined"
   />
 </template>
 
 <script setup lang="ts">
+import { isDeleteDateSoon } from '@base/utils/project';
+
 const props = defineProps< {
-  value: string|null,
+  value: string|'never'|null,
 }>();
+
+const apiSettings = useApiSettings();
+
+const isSoon = computed(() => isDeleteDateSoon(props.value));
 </script>