+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## v2026.50 - 2026-06-17
 * Move LLM configuration `AI_AGENT_MODELS` to web UI instead of env variables (https://docs.sysreptor.com/setup/configuration#llm-models)
 * Support multiple LLM models and model selection in UI
 * Rename configuration `AUTOMATICALLY_DELETE_ARCHIVED_PROJECTS_AFTER` to `AUTOMATICALLY_DELETE_PROJECTS_AFTER` (https://docs.sysreptor.com/setup/configuration#archiving)